🚚 refactor(curation): MonthNumberIconMapper 위치 이동 및 패키지 구조 정리
- **UI 매퍼 패키지 분리**
    - `MonthNumberIconMapper.kt` 파일을 `ui` 패키지에서 `ui.mapper` 하위 패키지로 이동
    - 파일 이동에 따라 패키지 선언을 `com.linku.curation.ui.mapper`로 변경

- **관련 의존성 업데이트**
    - `MonthlyCurationItem`에서 `resolveMonthNumberIcon` 참조 시 변경된 패키지 경로를 사용하도록 임포트 수정
