+++ feature/curation/src/main/java/com/linku/curation/ui/monthly/MonthlyCurationItem.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.tooling.preview.Preview
 import coil3.compose.AsyncImage
-import com.linku.curation.ui.resolveMonthNumberIcon
+import com.linku.curation.ui.mapper.resolveMonthNumberIcon
 import com.linku.design.theme.LinkuPreview
 import com.linku.design.theme.linkuColors
 import com.linku.design.util.scaler
